pyproject.toml
[tool.poetry]
name = "api"
version = "0.0.1"
description = "Task Manager API"
authors = ["kjplunkett <kris.j.plunkett@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
marshmallow-sqlalchemy = "^1.1.0"

[tool.poe.tasks]
lint = "flake8" # Lint with Flake8
format = "black ." # Format with Black
test = "pytest --cov=app tests/" # Test with pytest and pytest-cov
start = "flask --app app --debug run" # Start the API server

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
flake8 = "^7.1.1"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
